+++ src/messages/user_access_control/register_response_message.cpp
@@ -27,7 +27,7 @@ using namespace std;
 using namespace roa;
 
 template <bool UseJson>
-register_response_message<UseJson>::register_response_message(message_sender sender, int admin_status, int error_number, std::string error_str) noexcept
+register_response_message<UseJson>::register_response_message(message_sender sender, int8_t admin_status, int error_number, std::string error_str) noexcept
         : message<UseJson>(sender), admin_status(admin_status), error_number(error_number), error_str(error_str) {}
 
 template <bool UseJson>
